Implementing trapdoor and preimage as out-of-core algorithms
