+++ Cargo.toml
@@ -30,6 +30,7 @@ memmap2 = "0.9.5"
 tempfile = "3.19.1"
 sysinfo = "0.33.1"
 once_cell = "1.21.1"
+rand_distr = "0.5.1"
 # for now we put in test
